Refactor to monorepo structure
Use uv workspaces to manage multiple packages in a single repository.
Idea is to have 2 main packages with code they are dependent on living in the library package (divbase_lib).
