+++ pyproject.toml
@@ -1,21 +1,12 @@
 [project]
-name = "divbase_tools"
+name = "divbase_monorepo"
 version = "0.1.0"
-# Note pypi does not like AsciiDoc, may want to change to .md or .rst
-readme = { file = "README.adoc", content-type = "text/plain" }
+description = "Monorepo managed by uv"
 requires-python = ">=3.12"
-dependencies = [
-    "boto3>=1.38.15",
-    "minio==7.2.15",
-    "python-dotenv>=1.1.0",
-    "pyyaml>=6.0.2",
-    "typer>=0.15.4",
-    "pandas>=2.2.3",
+
+[tool.uv.workspace]
+members = [
+    "src/lib", 
+    "src/api", 
+    "src/cli"
 ]
-
-[build-system]
-requires = ["hatchling >= 1.26"]
-build-backend = "hatchling.build"
-
-[project.scripts]
-divbase-tools = "divbase_tools.__main__:app"
