+++ src/lib/divbase_lib/exceptions.py
