+++ src/lib/divbase_lib/utils.py
@@ -1,0 +1,2 @@
+def divbase_lib_says_hello():
+    print("Hello from divbase-lib!")
